refactor: LottoList를 LottoRepository로 이동
