+++ src/main/java/Main.java
@@ -4,7 +4,7 @@ import java.util.ArrayList;
 
 public class Main {
     public static void main(String[] args) {
-        LottoController controller = new LottoController(new ArrayList<>());
+        LottoController controller = new LottoController();
         controller.buildLotto();
         controller.checkWinNumber();
       }
